pyproject.toml
[project]
name = "doctorate-reader"
version = "0.1.0"
description = ""
authors = [
    {name = "Eduardo",email = "esanti@idealista.com"}
]
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "requests (>=2.32.5,<3.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "litellm (>=1.81.9,<2.0.0)",
    "fastapi (>=0.115.0,<1.0.0)",
    "uvicorn[standard] (>=0.30.0,<1.0.0)",
    "pyyaml (>=6.0,<7.0)",
]

[tool.poetry]
packages = [{include = "doctorate_reader", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
